tdf#114303 writerfilter: RTF import: fix default vert orient of shapes
The bugdoc is affected by the change of default vertical alignment, it
doesn't contain anything that would set the VertOrient property.

(regression from b26c0ad4c8de48e1b40a56cbcff07611a5966446)

Setting the AnchorType property before adding the shape is actually not
as effective as expected: it does cause the old default vertical
alignment to be used, but does not actually set the anchor type because
there is no anchor position and so SwXDrawPage::add resets the anchor
type to FLY_AT_PAGE; it's less confusing to set the VertOrient
property instead.

Change-Id: Ib0a71d20d479776ab5e4c21c3dd13c0c307a045e