Support for tcmalloc and jemalloc is long dead
When it was originally added to --with-alloc (tcmalloc:
f63ffaf3667e218be3538e84e6becc18d5de4334,
54efc5ed83890f0a1166ece3b1e4942bb7922ff7,
0abd71b44144220d2aeec58601929fd0cd1914d6
708b6d4246de915ac960c7b90ee8bb2d966c7d17 "INTEGRATION: CWS configure14: #i59160#
allow usage of tcmalloc and system-malloc"; jemalloc:
680b44c24eee6f8c579c043f44b47efc12150a5a "configure25: #i111045# jemalloc"), the
old build system's solenv/inc/unxlng*.mk added corresponding -l switches to
linker command lines, but it appears that was never carried over into the new
build system.

And even if the AC_CHECL_LIB calls in configure.ac prior to
9558e6639f44fcdb1fd06d9889ed53a489fc85ea "Ensure AC_CHECK_LIB/AC_SEARCH_LIBS do
not poison LIBS" would have extended LIBS with the correpsonding -l switches,
there appears to be no mechanism in the new build system that would have
depended on letting that have any effect on which alloc lib to link against.

So the only remaining effect of --with-alloc={tcmalloc,jemalloc} is to set
FORCE_SYSALLOC in Library_sal.mk, same as --with-alloc=system does.  So whoever
still uses the former configure switches should just switch to the latter.

(Also adapt the documentation of --with-alloc to reality, "internal" instead of
"oo".)

Change-Id: Iac0e302c5f9602daf5f8452a0fc1887dc55ee27b
Reviewed-on: https://gerrit.libreoffice.org/47178
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>