CppunitTest_sc_autoformatobj: fix read of value vs content of value
As far as I see the old qadevOOo/runner/lib/MultiPropertyTest.java
didn't assert the content of the property value (only avoided the set of
the property if the value was null or void), the new C++ version asserts
both the type and the value. Asserting the type is probably a good
thing, but asserting the value is problematic, at least the test failed
for me without these tweaks.

Change-Id: I688486025e6118ad4c8907de6cb2629dd008f8fd
Reviewed-on: https://gerrit.libreoffice.org/47475
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>