use the rect from all primitives when converting svg to Bitmap
In some cases we assume wrong size in svgio filter which we tried
to "fix" outside when we render svg to bitmap. This reverts the
"fix" and is now practically the same when rendering a svg
graphic (might be good to combine in the future). The fix for
svgio is fixed in a different commit.

Change-Id: I43ee35d3220b3a750b8f72f22085f6a34b9d26b8
Reviewed-on: https://gerrit.libreoffice.org/47497
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>