RotateFlyFrameFix: Add RotCenter info to ODF export
Since the WriterFlyFrames are implicitly rotated around
their center in the Writer UI this has to be added to the
transformation in ODF expressing this. The saved data
now contains the shape data in svg:x/y/width/height statements
(and thus back-compatible AFAP) and a draw:transform statement
(that is to be applied to the defined shape) containing the
rotation around the object's center.

Change-Id: Iccfc70924364f1d58995fce180bce9b6e7643ff4
Reviewed-on: https://gerrit.libreoffice.org/47719
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>